+++ manager.py
@@ -33,6 +33,27 @@ GCI_NUMBERS = re.compile(r'(_[0-9]{2})?\.gci', re.I)
 GCI_GLOB = '*.gci'
 
 
+def check_region(region: str):
+    """Checks and validates region from config.
+    Regions can only be 'EUR', 'JAP', or 'USA', as defined by
+    Dolphin Emulator.
+
+    Args:
+        region (str): the geographic region of the game's saves
+
+    Returns:
+        str: region, if valid
+
+    Raises:
+        Exception: if the config region is invalid
+
+    """
+    if region == 'EUR' or region == 'JAP' or region == 'USA':
+        return region
+    else:
+        raise Exception(f'{region} is an invalid region!')
+
+
 def convert_check_path(directory: str):
     """Expands tilde if present (convert). Checks the resulting path
     for existence.
@@ -54,7 +75,9 @@ def convert_check_path(directory: str):
         raise Exception(f'{directory} doesn\'t exist or isn\'t a directory')
 
 
-def check_file_conflicts(sub_dir: str, base_dir: str, card_slot: str):
+def check_file_conflicts(
+    sub_dir: str, base_dir: str, card_slot: str, region: str
+    ):
     """Checks for file conflicts. If conflicts exist, abort script.
     Otherwise, remove existing, valid symlinks.
 
@@ -66,6 +89,7 @@ def check_file_conflicts(sub_dir: str, base_dir: str, card_slot: str):
         sub_dir (str): the sub dir containing your saves elsewhere
         base_dir (str): the base dir created and used by Dolphin Emulator
         card_slot (str): either 'Card A' or 'Card B'
+        region (str): 'EUR', 'JAP', 'USA'
 
     Returns:
         bool: True, if files exist in `sub_dir` but not in
@@ -77,7 +101,7 @@ def check_file_conflicts(sub_dir: str, base_dir: str, card_slot: str):
     """
     non_empty = False
     failure = []
-    card_dir = base_dir / card_slot
+    card_dir = base_dir / 'GC' / region / card_slot
     for file in sub_dir.glob(GCI_GLOB):
         non_empty = True
         for g in card_dir.glob(GCI_NUMBERS.sub(GCI_GLOB, file.name)):
@@ -96,7 +120,7 @@ def check_file_conflicts(sub_dir: str, base_dir: str, card_slot: str):
             )
 
 
-def link_files(sub_dir: str, base_dir: str, card_slot: str):
+def link_files(sub_dir: str, base_dir: str, card_slot: str, region: str):
     """Link files from `sub_dir` into `base_dir` / `card_slot`.
     If `card_slot` doesn't exist, create it.
 
@@ -104,12 +128,13 @@ def link_files(sub_dir: str, base_dir: str, card_slot: str):
         sub_dir (str): the sub dir containing your saves elsewhere
         base_dir (str): the base dir created and used by Dolphin Emulator
         card_slot (str): either 'Card A' or 'Card B'
+        region (str): 'EUR', 'JAP', 'USA'
 
     Returns:
         bool: True
 
     """
-    card_dir = base_dir / card_slot
+    card_dir = base_dir / 'GC' / region / card_slot
     if not card_dir.exists():
         card_dir.mkdir(parents=True)
 
@@ -127,10 +152,11 @@ if __name__ == '__main__':
         conf = yaml.load(f, Loader=yaml.Loader)
     # Since `expanduser` doesn't care if '~' isn't present,
     # use it unconditionally
+    region = check_region(conf['region'])
     base_dir = convert_check_path(conf['base_dir'])
     save_dir = convert_check_path(conf['save_dir'])
     sub_dir = convert_check_path(save_dir / args.sub_dir)
     card_slot = f'Card {args.slot}'
     
-    if check_file_conflicts(sub_dir, base_dir, card_slot):
-        link_files(sub_dir, base_dir, card_slot)
+    if check_file_conflicts(sub_dir, base_dir, card_slot, region):
+        link_files(sub_dir, base_dir, card_slot, region)