+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0] - 2019-05-11
+## [1.0.1] - 2019-07-29
+### Added
+- Region checking for saves
+
+### Fixed
+- Typo in [`config.yaml.example`](config.yaml.example)
+- Use `'GC' / region` for Dolphin saves directory
+
+## [1.0] - 2019-07-16
 ### Added
 - Initial version